gl4es

+++ build_lwjgl.sh
@@ -14,11 +14,17 @@ cd ..
 rm -rf gl4es
 git clone --depth=1 https://github.com/ptitseb/gl4es.git
 cd gl4es
+sed -i 's/-fvisibility=hidden/-fvisibility=default/g' CMakeLists.txt #WE love ducttape and cableties
 mkdir build && cd build
-emcmake cmake .. -DNOX11=ON -DNOEGL=ON -DSTATICLIB=ON -DCMAKE_BUILD_TYPE=Release -DCMAKE_C_FLAGS="-fPIC -sRELOCATABLE=1 -sSHARED_MEMORY=1 -matomics -mbulk-memory"
+emcmake cmake .. -DNOX11=ON -DNOEGL=ON -DDEFAULT_ES=3 -DPLATFORM=EMSCRIPTEN -DSTATICLIB=ON -DCMAKE_BUILD_TYPE=Release -DCMAKE_C_FLAGS="-fPIC -sRELOCATABLE=1 -sSHARED_MEMORY=1 -matomics -mbulk-memory -pthread -fvisibility=default"
 emmake make -j$(nproc)
 cd ..
+mv lib/libGL.a lib/libGL4ES.a
 cd ..
+
+if [ "$1" = "gl" ]; then
+  exit
+fi
 
 rm -rf src
 wget https://github.com/LWJGL/lwjgl3/archive/refs/tags/$VER.zip
